Fix DeviceHandle::send_uapi_cmd
Previous implementation of api_exec was not flushing the writter after
writing to it. The new one does and hance the internal buffer of
BufWriter is cleared. That caused send_uapi_cmd to always return empty
string. With this fix we correctly return the external buffer to which
BufWritter is writing.