Make Persistent-Keepalive trigger handshakes
A Non-zero Persistent-Keepalive value triggers
handshake attempts even after 3xREJECT-AFTER-TIME.

Essentially, this ensures that handshake attempts are sent
to the offline peer all the time, every 5seconds.

This happens until the handshake is established or the
Persistent-Keepalive is set to zero. In which case after
the session expiration the handshake attempts are
continued for 3xREJECT-AFTER-TIME and then cease.

Whitepaper doesn't explicitly state that keepalives
should trigger handshakes but Linux kernel implementation
behaves like that. Also there's no penalty introduced in doing so
except for wasted resources while doing attempts.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>