Attempt handshakes if keepalive is set
In two separate cases it is possible enter
the state of no automatic handshake initiation
triggered by the keepalive which should happen.

This is done by having an expiration guard which is
a correct approach if there's no keepalive, but incorrect
if there's keepalive set. This makes it so that only
a packet which is sent to the tunnel for encapsulation
triggers the handshake ir receiving a trigger initiation
from the peer(which in VPN case should never happen).

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>