Upgrade wireguard-go to latest one
The latest wireguard-go performs benchmarks
at much better rate. It shows 0 packet drop and the
speed is about 5x of NepTUN in iperf download case.

This is very good since before this, wireguard-go
on my machine showed 2x slower speeds so we considered
it was already good enough and the packet drops were
present in both wireguard-go and NepTUN.

Now we can clealy see that packet drops should be zero
in this setup.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>