Enable 32bit target support
Use Mutex<u64> instead of AtomicU64 in rate limiter
to enable 32bit target support due to lack of 64bit
atomic support on 32bit targets.

An alternative is to use AtomicU32 which offers
degraded security.

Mutex approach is chosen because the values are only
used in the handshake phase, the hot path is avoided.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>